+++ packages/firestore/src/api/user_data_converter.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
+import { Timestamp } from '../api/timestamp';
 import { DatabaseId } from '../core/database_info';
-import { Timestamp } from '../core/timestamp';
 import { DocumentKey } from '../model/document_key';
 import { FieldValue, ObjectValue } from '../model/field_value';
 import {
@@ -526,6 +526,16 @@ export class UserDataConverter {
       return new StringValue(value);
     } else if (value instanceof Date) {
       return new TimestampValue(Timestamp.fromDate(value));
+    } else if (value instanceof Timestamp) {
+      // Firestore backend truncates precision down to microseconds. To ensure
+      // offline mode works the same with regards to truncation, perform the
+      // truncation immediately without waiting for the backend to do that.
+      return new TimestampValue(
+        new Timestamp(
+          value.seconds,
+          Math.floor(value.nanoseconds / 1000) * 1000
+        )
+      );
     } else if (value instanceof GeoPoint) {
       return new GeoPointValue(value);
     } else if (value instanceof Blob) {
@@ -602,6 +612,7 @@ function looksLikeJsonObject(input: AnyJs): boolean {
     input !== null &&
     !(input instanceof Array) &&
     !(input instanceof Date) &&
+    !(input instanceof Timestamp) &&
     !(input instanceof GeoPoint) &&
     !(input instanceof Blob) &&
     !(input instanceof DocumentKeyReference) &&