+++ packages/firestore/test/integration/api/fields.test.ts
@@ -18,12 +18,16 @@ import { expect } from 'chai';
 import firebase from '../util/firebase_export';
 import {
   apiDescribe,
+  DEFAULT_SETTINGS,
   toDataArray,
+  toIds,
   withTestCollection,
+  withTestCollectionSettings,
   withTestDoc
 } from '../util/helpers';
 
 const FieldPath = firebase.firestore.FieldPath;
+const Timestamp = firebase.firestore.Timestamp;
 
 apiDescribe('Nested Fields', persistence => {
   const testData = (n?: number): any => {
@@ -331,6 +335,95 @@ apiDescribe('Fields with special characters', persistence => {
         .then(() => coll.orderBy('field\\slash').get())
         .then(results => {
           expect(toDataArray(results)).to.deep.equal(expected);
+        });
+    });
+  });
+});
+
+apiDescribe('Timestamp Fields in snapshots', persistence => {
+  const testDataWithTimestamps = (ts: any): any => {
+    return { timestamp: ts, nested: { timestamp2: ts } };
+  };
+
+  it('are returned as native dates if timestampsInSnapshots is not set', () => {
+    const settings = { ...DEFAULT_SETTINGS };
+    settings['timestampsInSnapshots'] = false;
+
+    const timestamp = new Timestamp(100, 123456789);
+    const testDocs = { a: testDataWithTimestamps(timestamp) };
+    return withTestCollectionSettings(persistence, settings, testDocs, coll => {
+      return coll
+        .doc('a')
+        .get()
+        .then(docSnap => {
+          expect(docSnap.get('timestamp'))
+            .to.be.a('date')
+            .that.deep.equals(timestamp.toDate());
+          expect(docSnap.data()['timestamp'])
+            .to.be.a('date')
+            .that.deep.equals(timestamp.toDate());
+
+          expect(docSnap.get('nested.timestamp2'))
+            .to.be.a('date')
+            .that.deep.equals(timestamp.toDate());
+          expect(docSnap.data()['nested']['timestamp2'])
+            .to.be.a('date')
+            .that.deep.equals(timestamp.toDate());
+        });
+    });
+  });
+
+  it('are returned as Timestamps', () => {
+    expect(DEFAULT_SETTINGS['timestampsInSnapshots']).to.equal(true);
+
+    const timestamp = new Timestamp(100, 123456000);
+    // Timestamps are currently truncated to microseconds after being written to
+    // the database, so a truncated version of the timestamp is needed for
+    // comparisons.
+    const truncatedTimestamp = new Timestamp(
+      timestamp.seconds,
+      Math.floor(timestamp.nanoseconds / 1000) * 1000
+    );
+
+    return withTestDoc(persistence, doc => {
+      return doc
+        .set(testDataWithTimestamps(timestamp))
+        .then(() => doc.get())
+        .then(docSnap => {
+          expect(docSnap.get('timestamp'))
+            .to.be.an.instanceof(Timestamp)
+            .that.deep.equals(truncatedTimestamp);
+          expect(docSnap.data()['timestamp'])
+            .to.be.an.instanceof(Timestamp)
+            .that.deep.equals(truncatedTimestamp);
+
+          expect(docSnap.get('nested.timestamp2'))
+            .to.be.an.instanceof(Timestamp)
+            .that.deep.equals(truncatedTimestamp);
+          expect(docSnap.data()['nested']['timestamp2'])
+            .to.be.an.instanceof(Timestamp)
+            .that.deep.equals(truncatedTimestamp);
+        });
+    });
+  });
+
+  it('timestampsInSnapshots affects server timestamps', () => {
+    const settings = { ...DEFAULT_SETTINGS };
+    settings['timestampsInSnapshots'] = false;
+    const testDocs = {
+      a: { timestamp: firebase.firestore.FieldValue.serverTimestamp() }
+    };
+
+    return withTestCollectionSettings(persistence, settings, testDocs, coll => {
+      return coll
+        .doc('a')
+        .get()
+        .then(docSnap => {
+          expect(
+            docSnap.get('timestamp', {
+              serverTimestamps: 'estimate'
+            })
+          ).to.be.an.instanceof(Date);
         });
     });
   });